User submission form added for station info input by users
A user submission form, in a separate web page, was added to the map-display page in order to submit information to the website regarding specific information about each station. As the project evolves, the functionality will be implemented to enable automatic updates on the listings box, ensuring that users can promptly view any changes.